Move edge into its own module
